+++ src/Backend/SgateShopgatePlugin/Bootstrap.php
@@ -147,7 +147,7 @@ class Shopware_Plugins_Backend_SgateShopgatePlugin_Bootstrap extends Shopware_Co
 
     public function getVersion()
     {
-        return "2.9.84";
+        return "2.9.85";
     }
 
     public function getLabel()