+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Fixed
+- export of related and similar items
+
+## [2.9.85] - 2019-01-11
+### Added
+- support for the new Shopware PayPal Plugin
+
+### Fixed
+- potential security issue in redirect
 
 ## [2.9.84] - 2018-11-20
 ### Changed
@@ -769,7 +778,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - migrate plugin from Shopware 3.5.x to 4.0.x
 - use doctrine models
 
-[Unreleased]: https://github.com/shopgate/interface-shopware/compare/2.9.83...HEAD
+[Unreleased]: https://github.com/shopgate/interface-shopware/compare/2.9.85...HEAD
+[2.9.85]: https://github.com/shopgate/interface-shopware/compare/2.9.84...2.9.85
 [2.9.84]: https://github.com/shopgate/interface-shopware/compare/2.9.83...2.9.84
 [2.9.83]: https://github.com/shopgate/interface-shopware/compare/2.9.82...2.9.83
 [2.9.82]: https://github.com/shopgate/interface-shopware/compare/2.9.81...2.9.82